Move SelectionList components to SelectionList.tsx
